+++ src/client/activation/activationService.ts
@@ -103,7 +103,7 @@ export class LanguageServerExtensionActivationService implements IExtensionActiv
         //If an activated workspace folder was removed, dispose its activator
         const workspaceKeys = this.workspaceService.workspaceFolders!.map(workspaceFolder => this.getWorkspacePathKey(workspaceFolder.uri));
         const activatedWkspcKeys = Array.from(this.lsActivatedWorkspaces.keys());
-        const activatedWkspcFoldersRemoved = activatedWkspcKeys.filter(x => workspaceKeys.indexOf(x) < 0);
+        const activatedWkspcFoldersRemoved = activatedWkspcKeys.filter(item => workspaceKeys.indexOf(item) < 0);
         if (activatedWkspcFoldersRemoved.length > 0) {
             for (const folder of activatedWkspcFoldersRemoved) {
                 this.lsActivatedWorkspaces.get(folder).dispose();
